compiler: Remove special handling of non-extension options
Remove the special handling of option fields that are not extension
fields. They can be parsed the same way as extension fields are parsed.

One special case is the enum option `allow_aliases` - this option needs
to be known as we interpret the other options, otherwise the prototext
unmarshaler complains about aliased enum values. This occurs because the
resolver we use to unmarshal has no options set at all, whereas
previously the known field options were set in the resolver.

Update one test as the uninterpreted options is now populated with known
fields too.
